wip: refactor rapl limiter in line with amd
