+++ pkg/agent/sysadvisor/plugin/poweraware/component/capper/intel/rapllimiter_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package power
+package intel
 
 import (
 	"strings"
